refactor to remove flag
