+++ cmd/apply.go
@@ -9,7 +9,6 @@ import (
 func init() {
 	applyCmd.Flags().StringP("config", "c", "fogg.yml", "Use this to override the fogg config file.")
 	applyCmd.Flags().BoolP("upgrade", "u", false, "Use this when running a new version of fogg")
-	applyCmd.Flags().BoolP("oauth", "o", false, "Use this to clone Github remote modules using HTTP and an OAuth token stored in a FOGG_GITHUBTOKEN env variable")
 	rootCmd.AddCommand(applyCmd)
 }
 
@@ -37,11 +36,6 @@ var applyCmd = &cobra.Command{
 			return e
 		}
 
-		oauth, e := cmd.Flags().GetBool("oauth")
-		if e != nil {
-			return e
-		}
-
 		// check that we are at root of initialized git repo
 		openGitOrExit(fs)
 
@@ -54,7 +48,7 @@ var applyCmd = &cobra.Command{
 		}
 
 		// apply
-		e = apply.Apply(fs, config, templates.Templates, upgrade, oauth)
+		e = apply.Apply(fs, config, templates.Templates, upgrade)
 
 		return e
 	},
