+++ util/module_storage_test.go
@@ -4,10 +4,12 @@
 package util
 
 import (
+	"fmt"
 	"io/ioutil"
 	"os"
 	"testing"
 
+	getter "github.com/hashicorp/go-getter"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/require"
 )
@@ -34,7 +36,8 @@ func TestDownloadAndParseModule(t *testing.T) {
 	pwd, e := os.Getwd()
 	r.NoError(e)
 	fs := afero.NewBasePathFs(afero.NewOsFs(), pwd)
-	downloader := MakeDownloader("github.com/chanzuckerberg/fogg-test-module")
+	downloader, err := MakeDownloader("github.com/chanzuckerberg/fogg-test-module")
+	r.NoError(err)
 	c, e := downloader.DownloadAndParseModule(fs)
 	r.Nil(e)
 	r.NotNil(c)
@@ -42,4 +45,71 @@ func TestDownloadAndParseModule(t *testing.T) {
 	r.NotNil(c.Outputs)
 	r.Len(c.Variables, 2)
 	r.Len(c.Outputs, 2)
+}
+
+func TestMakeDownloader(t *testing.T) {
+	r := require.New(t)
+	creds := "REDACTED"
+	downloader, err := MakeDownloader("git@github.com:chanzuckerberg/shared-infra//terraform/modules/eks-airflow?ref=v0.80.0")
+	r.NoError(err)
+	r.Equal("git@github.com:chanzuckerberg/shared-infra//terraform/modules/eks-airflow?ref=v0.80.0", downloader.Source)
+
+	os.Setenv("FOGG_GITHUBTOKEN", creds)
+	defer os.Unsetenv("FOGG_GITHUBTOKEN")
+	downloader, err = MakeDownloader("git@github.com:chanzuckerberg/shared-infra//terraform/modules/eks-airflow?ref=v0.80.0")
+	r.NoError(err)
+	r.Equal(fmt.Sprintf("git::https://%s@github.com/chanzuckerberg/shared-infra//terraform/modules/eks-airflow?ref=v0.80.0", creds), downloader.Source)
+}
+
+func TestConvertSSHToHTTP(t *testing.T) {
+	r := require.New(t)
+
+	type test struct {
+		in  string
+		out string
+	}
+	creds := "REDACTED"
+	tests := func(token string) []test {
+		return []test{
+			{
+				in:  "git@github.com:chanzuckerberg/shared-infra//terraform/modules/eks-airflow?ref=v0.80.0",
+				out: fmt.Sprintf("git::https://%s@github.com/chanzuckerberg/shared-infra//terraform/modules/eks-airflow?ref=v0.80.0", token),
+			},
+		}
+	}(creds)
+	for _, test := range tests {
+		u := convertSSHToGithubHTTPURL(test.in, creds)
+		s, err := getter.Detect(u, creds, []getter.Detector{
+			&getter.GitHubDetector{},
+		})
+		r.NoError(err)
+		r.Equal(test.out, u)
+		r.Equal(test.out, s)
+	}
+}
+func TestConvertSSHToHTTPFail(t *testing.T) {
+	r := require.New(t)
+
+	type test struct {
+		in string
+	}
+
+	tests := func() []test {
+		return []test{
+			{
+				in: "github.com/scholzj/terraform-aws-vpc",
+			},
+			{
+				in: "terraform/modules/test",
+			},
+			{
+				in: "github.com/hashicorp/go-getter?ref=abcd12",
+			},
+		}
+	}()
+	for _, test := range tests {
+		s := convertSSHToGithubHTTPURL(test.in, "")
+		// make sure nothing changed in a failure case
+		r.Equal(test.in, s)
+	}
 }
